requirements.txt
   fastapi
   uvicorn
   sqlalchemy
   psycopg2-binary
   python-jose[cryptography]
   passlib[bcrypt]
   python-multipart
   pydantic
   python-dotenv
   pandas
   email-validator